fix: Apply clippy automatic fixes across workspace
Run cargo clippy --fix to automatically resolve clippy warnings.

Changes:
- Applied clippy's automatic fixes to 257 files
- Improvements include:
  - Simplified boolean expressions
  - Removed unnecessary clones and references
  - Used more idiomatic Rust patterns
  - Fixed redundant closures
  - Improved error handling patterns

All changes are clippy-suggested automatic fixes with no functional changes
to business logic.

+++ crates/reinhardt-template/crates/renderers/src/compression.rs
@@ -338,7 +338,7 @@ impl<R: Renderer> CompressionRenderer<R> {
 				use flate2::write::GzEncoder;
 
 				let mut encoder =
-					GzEncoder::new(Vec::new(), Compression::new(*level.min(&9) as u32));
+					GzEncoder::new(Vec::new(), Compression::new((*level.min(&9))));
 				encoder.write_all(data)?;
 				encoder
 					.finish()
